Current String class
Currently i have a reverse method and a SubString method. Since this allocates unmanaged memory when using either call Dispose method or declare a using statement. You can index the string with [] there is also implicit conversions you can assign a char* to a C obj or a C# String to a C obj. You can also assign a C obj to char* or char[]. when write the obj to the console it will treat it implicitly as a char[].